+++ lib/constants.h
@@ -3,8 +3,8 @@
 
 #include <cstdint>
 
-static const uint32_t ONE = (uint32_t) 1;
-static const uint32_t ZERO = (uint32_t) 0;
-static const uint32_t ALL_ONE = ZERO - ONE;
+static const uint64_t ONE = (uint64_t) 1;
+static const uint64_t ZERO = (uint64_t) 0;
+static const uint64_t ALL_ONE = ZERO - ONE;
 
 #endif //PRIMER_CONSTANTS_H
